feat: solve worker task in one file worker.js
