+++ src/wt/worker.js
@@ -1,16 +1,25 @@
-// чтобы увидеть работу этого worker, 
-// вызовите файл index.js с одним целым числом после имени файла (node .\wt\index.js 9)
+import path from 'path';
+import { argv } from 'process';
+import { fileURLToPath } from 'node:url';
+import { Worker, isMainThread, parentPort, workerData } from 'node:worker_threads';
 
-import { workerData, parentPort } from 'worker_threads';
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+const pathToWorker = path.join(__dirname, 'worker.js');
 
-// n should be received from main thread
+const num = argv[2];
+
 export const nthFibonacci = (n) => n < 2 ? n : nthFibonacci(n - 1) + nthFibonacci(n - 2);
 
 export const sendResult = () => {
-    // This function sends result of nthFibonacci computations to main thread
-    const result = nthFibonacci(workerData);
-    // parentPort
-    parentPort.postMessage(result);
-};
-
+  if (isMainThread) {
+    const worker = new Worker(pathToWorker, { workerData: String(nthFibonacci)});
+    worker.postMessage(num); 
+    worker.on('message', msg => console.log('Worker send message:', msg));    
+  } else {
+    parentPort.once('message', (message = 7) => {
+      const result = (eval(workerData))(message);  
+      parentPort.postMessage(result);
+    });
+  }
+}
 sendResult();
